regmapper: update all regmap tests
