+++ src/main/scala/amba/ahb/RegisterRouter.scala
@@ -115,6 +115,9 @@ trait HasAHBControlRegMap { this: RegisterRouter =>
     undefZero = undefZero,
     executable = executable)
 
+  // Backwards-compatibility default node accessor with no clock crossing
+  lazy val node: AHBRegisterNode = controlNode
+
   // Internally, this function should be used to populate the control port with registers
   protected def regmap(mapping: RegField.Map*) { controlNode.regmap(mapping:_*) }
 }
